Make Profiling_js.running mutable
Summary:
We constantly have to pass around `profiling` because it's immutable.
Every time we would time something, it would return a new record and we'd have
to propagate that. It was kind of a pain.

While mutability is usually pretty terrible, it's really useful to make
profiling and logging less of a burden.

Reviewed By: samwgoldman

Differential Revision: D5537185

fbshipit-source-id: a7af48684fb756112bb0904b6bcb24c7f1df0ea1