[Hack] Add ~weight to Measure.sample
Summary:
I've been meaning to do this for awhile. If we only call
`Measure.sample` 1 in 100 times, we can now do `~weight:100.` and each sample
will count like 100 samples.

I'm looking to keep track of how much time each error takes to generate. If,
after parsing for 10 seconds, we generate 4 errors, I want to say

  Measure.sample ~weight:4. "Time to error" 10.

The math for updating the count and the mean were pretty simple. I was less
confident for updating the variance sum, so I referred to
http://people.ds.cam.ac.uk/fanf2/hermes/doc/antiforgery/stats.pdf

Reviewed By: samwgoldman

Differential Revision: D5539302

fbshipit-source-id: 4d9d5ea6a9532caba28390bc652054c2b63c6e0a