Private class property type checking
Summary:
Adds typechecking for private class properties.

Whenever we enter a class body, we push a class binding to the environment
with the key ".class". There are new use types for setting and getting private
fields: SetPrivatePropT and GetPrivatePropT-- mirroring the already existing
SetPropT and GetPropT. Instead of looking at the instance itself, these new
private variats gather all the .class bindings in the environment into a list
and then look for it's class id. Once it is found, it takes the type of the property
from the private fields map. This allows us to have private fields with the same names
in the same scope but with different types depending on the subject of the access.

Reviewed By: samwgoldman

Differential Revision: D5576941

fbshipit-source-id: 375d2a164772221b97dd383199f96f2e96c4fbcf