Add React$ElementRef
Summary:
I will use this utility to implement refs in my next diff. It can also be used to improve typings for `React.render()` and other utilities which need the instance.

As a part of this change I altered the contract for `$JSXIntrinsics` to also require an `instance` property next to the `props` property.

Reviewed By: avikchaudhuri

Differential Revision: D5402443

fbshipit-source-id: 31ed4da91d5e7069c2472dff884fff8bcbd6ff69