Add type checking for special React key and ref props
Summary:
This specific implementation may break down with exact props. To fix that I'll diff out the `key` and `ref` props like we do with default props. If you have a better idea for an implementation let me know.

One of the first things I tried was reassigning config to:

```
config = {...config, ...{|key, ref|}}
```

Does not work for two reasons.

1. If `config` is not exact then all of its properties will become optional because Flow doesn't know if they are "own" or not. To fix this I would add some flag to `SpreadType` named something like `assume_own` for internal use only. This could also be used when I use a type spread to add `children`.
2. The type variable that `config` is set to from `get_props` has the props type as an upper bound. Not a lower bound. Type spreading straight up doesn't work because of that.

Reviewed By: samwgoldman

Differential Revision: D5407790

fbshipit-source-id: 342aa84ba0282eca80d1a595da69207d4732b305