Update SyntheticEvents
Summary: We aren't going to land D5417665 this round since the perf issues are going to take some time to resolve. However, I do want to land the update to the SyntheticEvent types that used to be in D5417665 because this update requires a codemod to add empty type arguments (`<>`) to all of the current SyntheticEvent references. Since we will be running codemods with the current batch it makes sense to do this as well.

Reviewed By: samwgoldman

Differential Revision: D5589699

fbshipit-source-id: dd2752938b078c2798582f3158959532b3741148