Add advanced HOC tests
Summary:
These tests are a sanity check for me because I want to recommend people to use the object spread type syntax for creating HOCs. (And I also want to make sure the other styles included with these tests are supported.)

These tests ensure that we don't regress this functionality which I'm going to put in the docs.

Reviewed By: samwgoldman

Differential Revision: D5597975

fbshipit-source-id: 743ec285b4dcfb999ffb0b3b19109b0120c210e0