add $CharSet type
Summary:
adds a `$CharSet<"abc">` type (where `"abc"` must be an inline string
literal) which acts like a union of all of the permutations of the power set of
`"abc"` (`"abc"`, `"ab"`, `"bc"`, `"a"`, `"b`", `"c"` and `""`).

the only use case I can think of is for the `flags` parameter of `RegExp`'s
constructor, where you can only pass strings containing the characters
"g", "u", "i", "m", "y" and "s" at most once. it's a syntax error if you mess
up, which is why we bother to catch it at all.

the "extra" info in the error message might be overkill. I included it because
it's stuff we know, and to try using `def_loc` to point to the original location
of the string literal rather than the callsite that errored.

Reviewed By: calebmer

Differential Revision: D5607790

fbshipit-source-id: aff7a13d045fbaf22e8971ed911c5f80d49eac12