[parser] fix parsing of computed and generator async methods
Summary:
we only parsed the `async` keyword on object initializers as a modifier
when it was followed by an identifier, as in `{ async foo() {} }`. computed
properties look like `{ async [foo]() {} }`, and async generators look like
`{ async *foo() {} }.

this fixes those cases, using the same heuristic that we use just below this
change to identify the following token.

Reviewed By: gabelevi

Differential Revision: D5601087

fbshipit-source-id: 041be6b2bbf63a2a8433857c30bc5a2f9a1485dc