[parser] fix frontmatter parsing in test262 runner
Summary:
using regexes to parse the frontmatter isn't great, but there are no
good yaml parser in ocaml.

to try to be as accurate as possible, I only allowed 0 or 1 leading spaces, so
that I could detect indented lines as 2+ spaces. but it turns out there are a
bunch of tests whose frontmatter is indented 2 spaces. this caused all the
matchers to fail, so we didn't know if those tests should expect parse errors,
or if they should only run in strict mode.

Reviewed By: samwgoldman

Differential Revision: D5619719

fbshipit-source-id: 48693aba7f78586ad33f7ece93b2f97fedc2fc40