[parser] fix yield expressions without arguments
Summary:
`YieldExpression` doesn't require an argument (https://tc39.github.io/ecma262/#prod-YieldExpression)

this lists all of the tokens that could not be the start of an `AssignmentExpression` that I can think of, and that cause test262 tests to fail.

note: we don't correctly handle some of the "early error" semantics, like the one that bans `YieldExpression` in `FormalParameters` (e.g. https://tc39.github.io/ecma262/#sec-generator-function-definitions-static-semantics-early-errors). some of the test262 tests were previously failing because they expected an argument to `yield` and thought `function *f(yield)` was an unexpected `)`. now they incorrectly parse "successfully" until we fix it so `YieldExpression` is forbidden. that explains the newly-failing test262 tests.

Reviewed By: samwgoldman

Differential Revision: D5624445

fbshipit-source-id: 8f174389b02ed5a98bc4baf4e91686ca02681e6e