Turn unused suppressions from errors into warnings
Summary:
Flow is intended to be a 0 error system. All errors should be fixed or
suppressed.

If you start off with 0 errors and then change a set of files, you should be
able to run `flow focus-check` on the changed files and see the same errors you
would see from `flow check`. This would allow CI systems to use `focus-check`
to verify that changes are safe.

In lazy mode, we sometime typecheck a dependency, without typechecking all of
that dependency's dependents. That means a suppression comment can show up in
that dependency and appear unused.

The easiest way to deal with this is to make these unused error suppression
errors into warnings. This way, we can choose to ignore these warnings when
they might be a side effect of lazy mode.

Reviewed By: avikchaudhuri

Differential Revision: D5605164

fbshipit-source-id: d3f09c7a23c45765dfd670c240fe7549ca304050