[PR] [travis] build with Trusty
Summary:
jobs are starting to run on their own with `dist: trusty`. we should pin to either `dist: precise` (old) or `dist: trusty` (new).

trusty seems to fail because (I assume) it has a newer `nvm`, which is not happy with the `PREFIX` that we export while setting up ocaml. I don't think `PREFIX` needs to be set anymore after it's installed, so let's try unsetting it.
Closes https://github.com/facebook/flow/pull/4606

Differential Revision: D5633024

Pulled By: mroch

fbshipit-source-id: fc06429cfe966676052e35df03ae6e893060d0b9