Fix polarity checking for property maps
Summary:
The polarity argument to `Flow_js.check_polarity` is a bit confusing, and
therefore was not always passed correctly.

In particular, recursive calls should use a value which is equivalent to
`Polarity.mult (p1, p2)` depending on the current value of the argument and the
polarity of the type being recursed into.

For example, function parameters have a negative polarity, so the recursive
calls pass `Polarity.inv polarity`, which is equivalent to
`Polarity.mult (polarity, Negative)`.

Conversely, function returns have a positive polarity, so the recursive call
passes `polarity`, which is equivalent to `Polarity.mult (polarity, Positive)`.

Finally, arrays are invariant, so the recursive call passes `Neutral`, which is
equivalent to `Polarity.mult (polarity, Neutral)`.

For property maps, we were not even considering the "current" polarity and we
only considered the polarity of the property itself. This caused us to
calculate the wrong result.

In particular, we previously allowed a type like this:

```
type O<-T> = ({-p: T}) => void;
```

But `T` actually appears in a *positive* position in the above type.

In general, I think it would be a lot clearer if the recursive calls passed in
the polarity of the type being recursed into, and a helper function dealt with
calling `Polarity.mult`.

Reviewed By: avikchaudhuri

Differential Revision: D5638469

fbshipit-source-id: 75eb384648056ad2241d97ac69a643905e349306