[Hack] Help `Worker.select` survive sigalrm
Summary:
On Linux and OS X, `Timer` and `Timeout` are based on sending sigalrm.
If `Unix.select` is blocking, then it will throw an `EINTR` exception. The best
thing to do in this case is just to catch that exception and resume the
`select`.

This comes up if you set a periodic timer on the master thread.

Reviewed By: alexchow

Differential Revision: D5473272

fbshipit-source-id: 436d1253a07bdb77455db9366528949c19472492