Include cpu info in flow_server_profile
Summary:
Honestly not sure how useful this will be, it was just easy to build.

As you may or may not know, running `flow --profile` will output a link to the
`flow_server_profiles` scuba table with samples from that run. It's useful for
dumping large data that wouldn't fit in `flow_server_events`. The recheck-time
vs file-count scatterplot uses this.

This diff adds a timer which fires every second and logs processor info to that
table.

Reviewed By: samwgoldman

Differential Revision: D5473737

fbshipit-source-id: 6668cc0576bb778a03e9cd3ca6cd60e845744374