Fix add-comments for adding JSX suppressions.
Summary:
The current mode of adding JSX suppression comments in `./tool` causes a lot of syntax errors and is visually unpleasing. At one point gabelevi mentioned to me that he was under the impression that:

```
<section>
  {/* Hello, world! */}
  <div />
</section>
```

Was invalid when in fact it is the only reason that [JSX](http://facebook.github.io/jsx/) has empty expression containers.

Reviewed By: gabelevi

Differential Revision: D5628726

fbshipit-source-id: 2af9112c17f8434cd1bcbd11ef77e354fdfa96a8