Implement mapped types as type destructors
Summary:
Before this change, mapped types were implemented as a def/use pair. The def
type used a `cont` to switch between upper bound and lower bound behavior.

This was a perfectly reasonable implementation, but using a type destructor is a
bit simpler. Since since mapped types only apply to annotations, which are 0->1,
the destructor's ReposUseT slingshot is sufficient to handle both lower bound
and upper bound behaviors.

One benefit of this approach is that EvalTs are cached, so we only need to
map over the type once. Before this change, everywhere a mapped type was used we
neeed to calculate the new props and create an entry in the context.

This probably won't lead to any big perf difference (fingers crossed though). I
just wanted to kill this unnecessary def type.

Reviewed By: calebmer

Differential Revision: D5632957

fbshipit-source-id: 1c018afc0d257c87ff7a06ec7aab36d6fa45c78f