[parser] fix patterns on LHS of for-in and for-of
Summary:
the LHS of for-in and for-of loops is either a variable/lexical
declaration, or a pattern. the spec is confusing because it says it should be a
`LeftHandSideExpression`, but then the [static semantics](https://tc39.github.io/ecma262/#sec-for-in-and-for-of-statements-static-semantics-early-errors) restrict it to patterns.

```
for ({x} in y) {}
for ([x] of y) {}
... etc
```

we were treating them as expressions instead of patterns.

Fixes https://github.com/facebook/flow/issues/3291

Reviewed By: gabelevi

Differential Revision: D5638265

fbshipit-source-id: c333a141ff9f8d4d99fb7849a83565f28c8ab7f5