Refactor React model and React.createElement()
Summary: This diff refactors `React.createElement()` so that we can remove `DefaultProps` as a type argument and instead infer it. This refactor also enables more features such as instances (needed for refs and the return type of `ReactDOM.render()`) and strongly typed children which will be coming in later diffs. I'll also be submitting diffs to implement `React.cloneElement()` and `React.createFactory()`.

Reviewed By: avikchaudhuri

Differential Revision: D5395485

fbshipit-source-id: d324c8304cadfad49aeb3f7abdb4729c0edd1f72