Persistent connection curr_lint_settings from the server env
Summary:
This fixes a crash we observed after releasing Flow 0.53. During a recheck,
when we go to update persistent clients, we check all new errors against the
current suppressions and lint_settings, to ensure we don't send suppressed
errors to clients.

However, we were starting with empty suppression and lint_settings maps each
recheck. This isn't correct, because an error in merge might reference a
location from a file that was previously checked, but not in the recheck set.

Instead, we should start with the set of suppressions and lint settings in the
server environment.

Reviewed By: gabelevi

Differential Revision: D5651286

fbshipit-source-id: 3d1706208ead287b96990191e9af2f35f998821a