[PR] serve_website.sh script to serve the website
Summary:
This is a rough copy of an internal script which we have to serve our website. I've updated it to work in our GitHub repo.

Some setup was necessary. I mainly followed the steps from https://help.github.com/articles/setting-up-your-github-pages-site-locally-with-jekyll/

```
$ \curl -sSL https://get.rvm.io | bash # installed rvm
$ rvm install 2.3.4
$ gem install bundler
$ cd website
$ yarn
$ bundle install
$ cd ..
$ ./scripts/serve_website.sh
```

It's also worth noting that this did NOT work with OCaml 4.04 and js_of_ocaml 2.8.4 but it did work with OCaml 4.03 and js_of_ocaml 2.8.1

Closes https://github.com/facebook/flow/pull/4632

Reviewed By: mroch

Differential Revision: D5651216

Pulled By: gabelevi

fbshipit-source-id: 035e71cea83f67814710f8aef634c15146c0b28d