[PR] Fix endless loop when running tests
Summary:
`runtests.sh` wrote its log file into the test directory. dfind is watching the test directory, and wakes up flow every time //any// file changes, even if it is then ignored by flow. every time flow wakes up, it sends any updates to its persistent connections and, critically, logs that it is doing so. this log change is picked up by dfind, making flow wake up and log, and on and on and on.

this changes to writing the log to the parent directory, which is also a temp dir created and cleaned up by the test runner.
Closes https://github.com/facebook/flow/pull/4636

Differential Revision: D5653835

Pulled By: mroch

fbshipit-source-id: 78a2b7c7797325dc31ad6bb0ff95df5fa4abb255