Add variant to distinguish between object spreads for annotations and values
Summary:
Originally, we wrote this code with only type spread in mind. With type spread,
whether to make the resulting object exact is a matter of annotation. With
values, the exactness of the result object depends on the exactness of the
input objects.

This change augments the representation of spreads to support both use cases. I
also intend to use this distinction to preserve variance annotations in type
spread, while creating invariant properties for values, which is more powerful.

Reviewed By: calebmer

Differential Revision: D5656539

fbshipit-source-id: 0aca40bab4f79a79773aa028b56333f9b3c00cda