Use sound spread mode to merge children in React.createElement calls
Summary:
Because the children mixin is exact, we get the expected behavior using the
standard merge mode for values.

React elements are almost always creating using JSX syntax, so the config
object will be exact. The children mixing object is constructed a few lines
above, and it is also always exact. Thus, the resulting object will also always
be exact.

This will change behavior for callers of React.createElement if the config
object is not literal (which is exact), but rather an annotation. This case is
rare, but the change in behavior is also a change from unsound to sound, which
I think is a good thing.

Reviewed By: calebmer

Differential Revision: D5656551

fbshipit-source-id: 844e58bf9eb4745d68842d7ef8385187b534a998