Remove now-unused basic merge mode, rename others
Summary:
The basic merge mode was only used to combine `{|children: T|}` with the React
element config object. Now that we use the standard spread implementation for
that, there is no remaining use for this mode. Hopefully we will not find a
need for an unsound merge mode in the future, either. If we do, this is easy
enough to add back in.

Reviewed By: calebmer

Differential Revision: D5656549

fbshipit-source-id: 171407b96c04a90b1eeb81c0bde190ad8321c426