[PR] [Docs] Explain function subtyping and variance
Summary:
I'm not an expert at type theory, but trying out the previous function subtyping example seemed to convince me that it was not correct because it explained that inputs and outputs should both be subtypes, whereas inputs should actually be a super type.

I'm also not an expert at technical writing, so please let me know (or just edit) if my tone and style don't fit with the rest of the docs.

I tried to add a light explanation of function subtyping that might help for simple cases, while still pointing to the Variance page for a more detailed discussion.

(Note that, in reading the Variance page, I notice that it mostly focuses on variance as applied to OO class hierarchies. I think there might be room for a simpler explanation of function input/output types. There is also probably a need for some examples of just functions with complex cases: arguments that are callbacks, whose arguments are in output position and therefore covariant, arguments that are written to, and therefore invariant, etc.)
Closes https://github.com/facebook/flow/pull/4635

Differential Revision: D5660256

Pulled By: jbrown215

fbshipit-source-id: bfe0b6c447f61cc34c2eb01c637ca7a349259ec3