Persistent connection curr_severity_cover from the server env
Summary:
This fixes a crash we observed after releasing Flow 0.53. During a recheck,
when we go to update persistent clients, we check all new errors against the
current suppressions and severity cover, to ensure we don't send suppressed
errors to clients.

However, we were starting with empty suppression and severity cover maps each
recheck. This isn't correct, because an error in merge might reference a
location from a file that was previously checked, but not in the recheck set.

Instead, we should start with the set of suppressions and severity cover in the
server environment.

Reviewed By: nmote

Differential Revision: D5655097

fbshipit-source-id: a3c0552ed91b3a876712622c810184acfb88d262