Fix exported type of React.StatelessFunctionalComponent
I don't know what the C type variable was supposed to model, but it was seemingly causing a lot of problems. In the thought that it might have been for the context types, I've also added the contextTypes as an optional value on a stateless functional component following the pattern currently used for full Components.