Continue Flow checking after EIncompatibleWithExact
Summary:
This is a small UX thing. Often `EIncompatibleWithExact` is a symptom of an underlying incompatibility between types but is not the root cause. For example: `number ~> {||}`. The root cause of this error is that number is not an object! However, the user is only shown `EIncompatibleWithExact` as type checking stops.

Noticed this as I was investigating errors added in `www` after my React diffs. Turning `React$Element` into an exact type meant I lost a lot of valuable information.

Reviewed By: samwgoldman

Differential Revision: D5475999

fbshipit-source-id: e4de2c2bc430c395bd547ab5cec6097e5ea8d790