add FLOW_TEMP_DIR env
Summary:
lets you customize flow's temp directory via an environment variable
in addition to the `--temp-dir` flag, as in `FLOW_TEMP_DIR=/tmp/flow-dev flow`.

this is useful during development, when we often want to run a dev build on the
same directory as the release build, but normally the pid/lock/log files would
step on each other. I have a complex wrapper script that tries to insert
`--temp-dir` in the right place, but this will make it super simple.

I also use this in the test runner, so that the temp files are stored
in the temp dir the test runner creates, instead of polluting /tmp/flow/.

Reviewed By: samwgoldman

Differential Revision: D5659481

fbshipit-source-id: 124c75a4a6774aebb4b159fe7f9316a8665028ed