[parser] `super` must always be a member or call expression
Summary:
`super` is a reserved word that can only appear in a [SuperCall](https://tc39.github.io/ecma262/#prod-SuperCall)
or a [SuperProperty](https://tc39.github.io/ecma262/#prod-SuperProperty).

this diff removes it from `primary` since it isn't a `PrimaryExpression`, and
calls `call` on it, which expects it to be part of a member expression or a call
expression.

Reviewed By: samwgoldman

Differential Revision: D5646321

fbshipit-source-id: fec50fc86de650eaadafe400463734a90cf89497