[PR] Fix exported type of React.StatelessFunctionalComponent
Summary:
Resolves #4650.

I don't know if the `C` type variable was intended to model something else, but it was not used and seemingly causing a lot of problems. Removing it made things work as (I, at least) expected.

I also noticed that the tests previously did not contain any explicit usages of `StatelessFunctionalComponent`, which is presumably how this wasn't noticed. So I added a new test using that type, which passes only after this commit.

(I've also added the contextTypes as an optional value on a stateless functional component following the pattern currently used for full Components. That could be committed separately, but I thought it was worth adding while I was in there...)
Closes https://github.com/facebook/flow/pull/4652

Reviewed By: samwgoldman

Differential Revision: D5679949

Pulled By: calebmer

fbshipit-source-id: 40a0b8fcedcbf873293254aead106f839f192d68