[parser] use cover grammar to parse the LHS of assignments
Summary:
`ObjectLiteral` contains a `PropertyName: CoverInitializedName`
production which allows `{ foo = 123 }`, to make it a cover grammar for
`ObjectAssignmentPattern`. then it forbids `CoverInitializedName` unless the
goal is a pattern (e.g. the LHS of an `AssignmentExpression`). see [12.2.6.1](https://tc39.github.io/ecma262/#sec-object-initializer-static-semantics-early-errors).

this diff makes `primary` return an `object_cover` variant instead of an
`Ast.Expression.t` directly. `object_cover` differentiates `ObjectLiteral`
from other expressions, and provides a place to remember the location of
assignment operators.

as it works its way back up the tree, we convert it back to an expression if
it's used in an expression context. `assignment` is looking for a pattern on the
LHS, so it calls `condition_cover` which drills down to `primary_cover`. if a
`Cover_object` is returned from `condition_cover`, then we can interpret it as a
pattern (allowing `CoverInitializedName`) instead.

this diff does //not// add support for `CoverInitializedName` yet. it just
provides plumbing. a followup diff will make `object_initializer` support
`key = value`, but also return the location of each `=` and pass those locs in
`Cover_object`. if it reaches an expression location, we throw an
"Unexpected token `=`" error at those locations.

Reviewed By: samwgoldman

Differential Revision: D5675154

fbshipit-source-id: f5b8d010f5b563e614d016f50f75cefc15bc54e1