[parser] support defaults in ObjectAssignmentPattern
Summary:
in a pattern, `({ foo = 123 } = {})` assigns `123` to `foo`. when
parsing an object literal, this diff allows `{ foo = 123 }` as shorthand for
`{ foo: foo = 123 }`. this is allowed by [//CoverInitializedName//](https://tc39.github.io/ecma262/#prod-CoverInitializedName)
but then forbidden by static semantics unless we're in a pattern context.

D5675154 added plumbing for keeping track of the location of the `=`. we pass
the location up the callstack until we can decide whether we're in a pattern or
expression. if it's a pattern, great. if it's an expression then we throw
"Unexpected token `=`" errors at those locations.

Reviewed By: samwgoldman

Differential Revision: D5675240

fbshipit-source-id: 9fad24f39477457e7c3aafab1a56ec75377696fe