Merge master cx before local inference
Summary:
Before this change, any builtin lookup would be deferred on a tvar until after
local inference was complete. We would finally merge the master cx, which would
resolve a cascade of constraints.

This makes it possible to remove some crufty code that we used to terminate
prototype lookups, which would normally need to wait until Object.prototype was
available. In general, we were careful to avoid building deferred constraints
based on Object.prototype and Function.prototype.

Reviewed By: avikchaudhuri

Differential Revision: D5619326

fbshipit-source-id: 272b70f53e1f76f38d7840aead8ca11608a80032