Seal object rest and make exact if possible
Summary:
Before this change, an object created through a rest element in a destructuring
pattern would always be unsealed, which is inconvenient now that we have exact
object annotations. For example, the following would not type check:

```
var a = {p: 0, q: 0};
var {p, ...b} = a;
(b: {|q: number|});
```

This would error because `b` was unsealed. Because unsealed objects can be
widened by arbitrary writes, unsealed objects are never compatible with exact
annotations.

While I was in here, I noticed that we don't deal with unsealed objects in a
great way. I fixed one particularly ugly issue where shadow properties were
copied over, meaning writes to the rest result reached reads from the original
object.

As far as a complete design, I have one idea, but implementation of that is
left for future work.

Reviewed By: calebmer

Differential Revision: D5667092

fbshipit-source-id: 4de7f96e8162d6ccec081c6157f62ebcdb09d4d1