process commands during recheck
Summary:
Currently the Flow server's main loop listens for either file system changes or
commands, responding synchronously by either triggering a recheck or processing
the commands. This means that during a long recheck, commands must wait to be
processed. Overall this can make Flow-based code intelligence tooling provided
through an IDE go AWOL for a while after saving edits.

This diff passes a handler to the recheck function that can listen for and
process commands (with timeout=0, which means that it returns immediately if
there's no outstanding request). The handler is called every time a worker is
done checking a batch of files (actually, it is called as part of the callback
that streams errors).

Reviewed By: samwgoldman

Differential Revision: D5590460

fbshipit-source-id: 521a1b90750805284a5a1f60eeb1a9914f95e416