Enable phantom types by comparing type args only in type apps with the same polymorphic type
Summary:
The following currently does not produce an error:

```
type T<Phantom> = any;
type A = T<number>;
type B = T<string>;

declare var a: A;
(a: B);
```

https://flow.org/try/#0C4TwDgpgBAKgPABQBYEMB2wD2BbAfFAXinRAG4AoUSKAQUNjjQFdsAjCAJ1wqugCF68AM7AOASzQBzbuXIATCAGMANig7QAbmuIAuWhQAUKPXwCUFIA

However, I believe that it should. Looking at the code it appears like the intention was to make this an error by sending `TypeAppT (c1, _) ~> TypeAppT (c2, _)` when `c1` and `c2` are the same to `TypeAppVarianceCheckT`. However, to check if `c1` and `c2` are the same we performed a referential equality check (`c1 = c2`).

This does not catch every case where `c1` and `c2` are the same. Notably it doesn't catch in the above case where the `PolyT` for `T` is repositioned in such a way that `c1` and `c2` for `A` and `B` respectively are unequal. Also, if you move `T` to the bottom of the file, say:

```
type A = T<number>;
type B = T<string>;

declare var a: A;
(a: B);

type T<Phantom> = any;
```

Then `c1` and `c2` for `A` and `B` respectively will be two different tvars.

This diff replaces the referential equality check (`c1 = c2`) with a step further. We concretize both `c1` and `c2` then check the referential equality of the `PolyT` def types.

I discovered this bug (at least I believe it's a bug, correct me if I'm wrong) when exploring how to type Relay `createFragmentContainer()`. Relevant to this diff is typings that look like:

```
type Fragment<Ref, Shape> = Shape;

declare opaque type ViewerFragmentRef;
type ViewerFragment = Fragment<ViewerFragmentRef, number>;
declare var viewerFragment: ViewerFragment;
declare var viewerFragmentRef: ViewerFragmentRef;

({
  viewer: viewerFragmentRef,
}: $ObjMap<
  {
    viewer: ViewerFragment,
  },
  <Ref, _Fragment: Fragment<Ref, mixed>>(_Fragment) => Ref,
>);
```

`Fragment` has a phantom type argument `Ref`. When we Flow `Fragment<ViewerFragmentRef, number>` into `_Fragment` we want `ViewerFragmentRef` to reach `Ref`. If we expand the type apps then `ViewerFragmentRef` will be ignored and `Ref` will only have an upper bound of `mixed`.

Reviewed By: samwgoldman

Differential Revision: D5648147

fbshipit-source-id: 566c266b57271d2fc8606a85abffc3c913f242af