[parser] support destructuring defaults in array assignments
Summary:
if an object destructuring is nested inside an array destructuring, we
need the array to keep track of the object's `=` until we decide whether the
array is a pattern or expression. this diff changes from `Cover_object` to
`Cover_patt` which wraps an expression and keeps track of the `=` locs.

Reviewed By: samwgoldman

Differential Revision: D5682629

fbshipit-source-id: b52462b106039bfabac6004a8325ce86c3463896