[parser] minor for-loop parsing refactor
Summary:
rather than creating `For.InitExpression` and `For.InitDeclaration`s
and then recasting them into `ForOf.LeftDeclaration`, etc., this creates a
temporary variant and then converts them into the proper type in all 3 types of
for loop (for, for-in, for-of) instead of just the latter two.

the current behavior relies on for-in and for-of LHSs being covered by for's
init, but that's not actually the case. this diff will make a followup diff
simpler, since I can change the type of `For_expression`. I can't change it now
since it has to be a valid for-loop AST right now.

Reviewed By: samwgoldman

Differential Revision: D5681753

fbshipit-source-id: 22791d2b4e918dd084499de6f82ec42a5ffc9026