[parser] fix `let [` negative lookahead in ExpressionStatement
Summary:
`ExpressionStatement` has some negative lookaheads, which we implement
in `expression_statement`, but they weren't being hit because
`Peek.is_identifier` is very broad and includes `T_LET`, and the
`_ when Peek.is_identifier env` case is above the call to
`expression_statement`.

this is the second time this has happened. I regret making
`expression_statement` its own function. this diff merges it back into
`statement` and orders the patterns properly.

while I was there, I also made a specialized error since `let` on its own is
allowed, so it's confusing to point just at `let`, saying it's unexpected, when
it's really the `[` (or the combo of them) that's unexpected.

Reviewed By: avikchaudhuri

Differential Revision: D5684711

fbshipit-source-id: 8f257d2bcd202f2592564fc4382f1a88d15e60b5