Method shorthand properties in object literals should be invariant
Summary:
Originally I thought we needed to do this, but an object with an invariant
property is more flexible than an object with a covariant property. That is,
the literal `{m() {}}` is still a subtype of of the type `{m(): void}`.

This does cause some churn in type-at-pos and gen-flow-files, but the churn is
accurate, because we should represent the types of these literals as invariant.

Reviewed By: calebmer

Differential Revision: D5684003

fbshipit-source-id: 6430fea2155f0cc722b9456783985e99c8a77f40