Don't copy/restore dependency sig contexts
Summary:
Sig contexts are copied quite a lot in workers. The heap itself has a
worker-local cache. On top of that, we cached sig cxs in a hash table.

The latter cache was important because merge used to involve mutations to the
graph in the sig cxs. However, now that we perform context optimization, the sig
cxs become effectively read-only.

Consider a diamond-shaped dependency graph, like so:

```
  A
 / \
B   C
 \ /
  D
```

If B and C include parts of A in their exports, it's important that they agree,
so that D will see a consistent view of those parts of A, whether they came from
B or C.

However, because A is effectively "closed" by context optimization, I can't
think of any way that B or C could export a signature that is inconsistent with
A.

Reviewed By: avikchaudhuri

Differential Revision: D5696578

fbshipit-source-id: 244c1fae3fb3df90aef65a0bee31bfe06503d2fd