[parser] recover gracefully if expected .tree.json file is malformed
Summary:
when I'm rebasing and get merge conflicts, I often leave the conflict
markers and re-record the expected output.

it was failing because <<<<<< is not valid JSON and we raised on JSON parse
errors. now, we fail the test but print the parse error. `-r` is able to
re-record now that it doesn't crash.

Reviewed By: samwgoldman

Differential Revision: D5705310

fbshipit-source-id: c13724a8c775ded54bf54d0c7882214e34858286