[parser] remove raw from BooleanLiteral ast
Summary:
`true` and `false` are the only character sequences that represent
boolean literals (the only other potential option would be if escape codes were
valid, but they aren't), so there's no need to carry around the raw value since
it can be computed on demand.

came across this while trying to kill a `Peek.value` call.

Reviewed By: avikchaudhuri

Differential Revision: D5705887

fbshipit-source-id: 0ea1425cc2379dbcdb36b19e259b3a7f9dbe7243