Add how to extend the constructor and prototype
Addressing issue #4717.
These changes make it clear how to extend the prototype and constructor of a class. Methods added in the class declaration do not need to be annotated. At this time, methods and properties defined on the constructor (static) and on the prototype need to be annotated in the class definition. This is probably a best practice since, if the class is modified in a location not local to the definition we will want the type checker to notify us.