[parser] refactor IdentifierName parsing
Summary:
an //Identifier// is an //IdentifierName// that is not a //ReservedWord//
(which includes keywords). `identifier_or_reserved_keyword` essentially implements
this, returning an `Identifier.t` and an optional error that you should throw if
you're looking for an //Identifier// and not an //IdentifierName//.

we are inconsistent about handling the errors properly. sometimes they are
ignored when they should raise instead.

I think the current implementation of `identifier_or_reserved_keyword` is poorly
named and a bit inscrutible, so this diff adds `identifier_name` which takes a
token and returns `Some name` if `name` is a valid //IdentifierName//, or `None`. So it
returns `None` for `T_RCURLY`, for example. Callers should only call this if
they're looking for an //IdentifierName//, and call `Parse.identifier` if they
want an //Identifier//.

Reviewed By: samwgoldman

Differential Revision: D5709853

fbshipit-source-id: 13fe6c1f9b9c4d4aed819e52d599d626637e11b2