[parser] kill Peek.value
Summary:
some tokens have variable values, which they carry around with them
(e.g. `T_IDENTIFIER`). most are static, like `T_RCURLY`.

we only need the value of an arbitrary token for the token stream in `flow ast` and
for printing `Unexpected token <value>` errors.

otherwise, matching the value of the token instead of the token's variant is
sloppy and will lead to problems with Unicode, where there are many possible
encodings of the same value.

Reviewed By: samwgoldman

Differential Revision: D5715114

fbshipit-source-id: 91d41bf7d70c743259356869c842fcf0a398151e